fixing bugs for final version
