+++ package.json
@@ -7,10 +7,8 @@
         "@testing-library/jest-dom": "^5.16.5",
         "@testing-library/react": "^13.4.0",
         "@testing-library/user-event": "^13.5.0",
-        "nanoid": "^4.0.2",
         "react": "^18.2.0",
         "react-dom": "^18.2.0",
-        "react-icons": "^4.8.0",
         "react-router-dom": "^6.9.0",
         "react-scripts": "5.0.1",
         "react-slideshow-image": "^4.2.1",
