+++ src/styles/layout/_cards.scss
@@ -5,6 +5,9 @@
     padding: $layout-mobile-padding;
     gap: 20px;
     margin-block: $layout-mobile-padding;
+    @include tablette-only {
+        gap: 30px;
+    }
     @include desktop-only {
         margin-block: $layout-desktop-padding;
         grid-template-columns: 1fr 1fr 1fr;
@@ -44,13 +47,17 @@
             color: $color-white;
             padding: 20px;
         }
-        &__gradient{
+        &__gradient {
             border-radius: 10px;
             position: absolute;
-            background: linear-gradient(180deg, rgba(255, 255, 255, 0) 0%, rgb(0, 0, 0, 0.7) 100%);
-            height:100%;;
-            width:100%;
-            top:0%;
+            background: linear-gradient(
+                180deg,
+                rgba(255, 255, 255, 0) 0%,
+                rgb(0, 0, 0, 0.7) 100%
+            );
+            height: 100%;
+            width: 100%;
+            top: 0%;
         }
     }
 }
