+++ src/styles/layout/_slides.scss
@@ -46,6 +46,9 @@
         @include mobile-only {
             font-size: 1em;
         }
+        @include tablette-only {
+            font-size: 1.5em;
+        }
     }
     &__bullet {
         opacity: 70%;
@@ -65,6 +68,9 @@
             object-fit: cover;
             @include mobile-only {
                 height: 255px;
+            }
+            @include tablette-only {
+                height: 300px;
             }
         }
     }
