+++ src/styles/utils/_mixins.scss
@@ -1,10 +1,16 @@
-@mixin desktop-only{
-    @media screen and (min-width: map-get($breakpoints, mobile)){
+@mixin desktop-only {
+    @media screen and (min-width: map-get($breakpoints, tablet)) {
         @content;
     }
 }
-@mixin mobile-only{
-    @media screen and (max-width: map-get($breakpoints, mobile)){
+@mixin mobile-only {
+    @media screen and (max-width: map-get($breakpoints, mobile)) {
+        @content;
+    }
+}
+
+@mixin tablette-only {
+    @media screen and (min-width: map-get($breakpoints, mobile)) and (max-width: map-get($breakpoints, tablet)) {
         @content;
     }
 }
